fix: Text matchers unable to capture execution context (#28)
# What?

For some reason `toHaveText` and `toContainText` return an empty stack
when imported from https://jslib.k6.io, causing the
`captureExecutionContext` to fail. Adding a console.log for the stack
showed that it captured the stack correctly when run locally, but the
same console.log returns an empty stack when imported from a server.

It seems to have been caused by using a higher-order function to define
both matchers. This PR refactors those matchers to call a helper
function directly.

# How to test

It is not enough to import `dist/index.js` locally. For whatever reason
that works as expected.

The way I've tested it is by uploading the file directly to S3 (I used
the app.staging.k6.io bucket) and then imported it in my script from
there.